refactor : shift possible proposal visibility test from logged-in to non-logged-in
